+++ plugins/crisp/src/editor.rs
@@ -31,7 +31,7 @@ impl Model for Data {}
 
 // Makes sense to also define this here, makes it a bit easier to keep track of
 pub(crate) fn default_state() -> Arc<ViziaState> {
-    ViziaState::from_size(400, 390, false)
+    ViziaState::new(|| (400, 390))
 }
 
 pub(crate) fn create(