+++ plugins/examples/gain_gui_vizia/src/editor.rs
@@ -19,7 +19,7 @@ impl Model for Data {}
 
 // Makes sense to also define this here, makes it a bit easier to keep track of
 pub(crate) fn default_state() -> Arc<ViziaState> {
-    ViziaState::from_size(200, 150, false)
+    ViziaState::new(|| (200, 150))
 }
 
 pub(crate) fn create(