+++ plugins/spectral_compressor/src/editor.rs
@@ -36,7 +36,7 @@ impl Model for Data {}
 
 // Makes sense to also define this here, makes it a bit easier to keep track of
 pub(crate) fn default_state() -> Arc<ViziaState> {
-    ViziaState::from_size(680, 535, false)
+    ViziaState::new(|| (680, 535))
 }
 
 pub(crate) fn create(